WIP Add Page-Breaking Functionality for New Filter
This commit adds the functionality used in the `Technologies` and
`Languages` filter to the new `Tools Filter`.

This commit breaks the page due an incompatibility between the current
filter implementation and the data format for `Tools`. The
implementation expects a YAML array, but the data is currently stored
as a YAML string of comma-separated values.